requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12

# LangChain (nouvelle API, non-dépréciée)
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-ollama==0.2.0
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2

# Vector store + embeddings
faiss-cpu==1.9.0
sentence-transformers==3.2.1

# Supabase
supabase==2.9.1
postgrest==0.17.2

# --- Parseurs de documents pour la branche COURSE ---
# PDF
pypdf==5.1.0
pdfplumber==0.11.4
# Word
python-docx==1.1.2
# PowerPoint
python-pptx==1.0.2
# OCR (images scannées + PDF scannés)
# Note : tesseract doit être installé au niveau OS
# Linux:   sudo apt install tesseract-ocr tesseract-ocr-ara tesseract-ocr-fra poppler-utils
# Windows: https://github.com/UB-Mannheim/tesseract/wiki + poppler-windows
# macOS:   brew install tesseract tesseract-lang poppler
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0

# Utilitaires
langdetect==1.0.9
cachetools==5.5.0
